requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-magic
telegraph
tenacity
uvloop
pyrogram
tgcrypto
beautifulsoup4
colorama
construct
cryptography
cssutils
ffmpy
future
HTMLParser
httpx
inquirer
isodate
natsort
nltk
pproxy
prettytable
protobuf==3.20.1
pycountry
pycryptodome
pycryptodomex
PyJWT
pymediainfo
pysrt
Requests
six
svtplay_dl
termcolor
titlecase
tldextract
toml
tqdm
Unidecode
urllib3
wcwidth
xmltodict
yarl
hpack
hyper
hyperframe
h2
expiringdict
